+++ HISTORY.md
@@ -1,3 +1,8 @@
+# TileDB-Py 0.9.6 Release Notes
+
+## Bug Fixes
+* Correct listing of consolidated fragments to vacuum in the Fragment Info API by deprecating `FragmentInfoList.to_vacuum_uri`, `FragmentInfoList.to_vacuum_num`, `FragmentInfo.to_vacuum_uri`, and `FragmentInfo.to_vacuum_num` and replacing with `FragmentInfoList.to_vacuum` [#650](https://github.com/TileDB-Inc/TileDB-Py/pull/650)
+
 # TileDB-Py 0.9.5 Release Notes
 
 ## TileDB Embedded updates: